refactor: 이벤트 신청 mutation 페이지에서 전달
