+++ src/pages/EventPages/EventDetailPage/EventDetailPage.tsx
@@ -1,4 +1,5 @@
 import { Flex, useDisclosure } from '@chakra-ui/react';
+import { useQueryClient } from '@tanstack/react-query';
 import { Suspense } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { Spinner } from '~/components/atoms/Spinner';
@@ -9,6 +10,8 @@ import ResumeSelect from '~/components/organisms/ResumeSelect/ResumeSelect';
 import { appPaths } from '~/config/paths';
 import useUser from '~/hooks/useUser';
 import { useGetEventDetail } from '~/queries/event/details/useGetEventDetail';
+import { eventKeys } from '~/queries/event/eventKeys.const';
+import usePostEventApply from '~/queries/event/usePostEventApply';
 import { useGetMentorDetail } from '~/queries/user/details/useGetMentorDetail';
 
 const EventDetailPage = () => {
@@ -16,12 +19,14 @@ const EventDetailPage = () => {
 
   const navigate = useNavigate();
 
-  const { eventId = '' } = useParams();
+  const { resumeId = '', eventId = '' } = useParams();
 
   const { data: event } = useGetEventDetail({ eventId });
   const { data: mentor } = useGetMentorDetail({ mentorId: String(event.mentorId) });
+  const { mutate: postEventApplyMutate } = usePostEventApply();
 
   const { isOpen, onOpen, onClose } = useDisclosure();
+  const queryClient = useQueryClient();
 
   return (
     <>
@@ -30,7 +35,20 @@ const EventDetailPage = () => {
         onClose={onClose}
       >
         <Suspense fallback={<Spinner />}>
-          <ResumeSelect onCancel={onClose} />
+          <ResumeSelect
+            onCancel={onClose}
+            onSubmit={() => {
+              postEventApplyMutate(
+                { resumeId: parseInt(resumeId), eventId },
+                {
+                  onSettled: () => {
+                    onClose();
+                    queryClient.refetchQueries({ queryKey: eventKeys.getEventDetail(eventId) });
+                  },
+                },
+              );
+            }}
+          />
         </Suspense>
       </Modal>
       <Flex
